custom-elements/creating-and-passing-registries/existing-registry-test-003.html

<!DOCTYPE html>
<html>
<head>
<title>When creating an import, use the registry of the master document</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="When creating an import, use the registry of the master document">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#creating-and-passing-registries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

testInIFrame('../resources/import-master.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    
    var link = doc.querySelector('link[rel=import]');
    
    var doc2 = link.import;
    
    var name = 'x-frame';
    
    doc.register(name);
    
    assert_throws(null, function () {
        doc2.register(name); 
    }, 'Exception should be throw if element with the same type is already registered');
    
}, 'Registry of the import should share master document registry');


testInIFrame('../resources/import-master.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    
    var link = doc.querySelector('link[rel=import]');
    
    var doc2 = link.import;
    
    var name = 'x-frame';
    
    doc.register(name);    
    
    assert_throws(null, function () {
        doc2.register(name); 
    }, 'Exception should be throw if element with the same type is already registered');
    
    var doc3 = newHTMLDocument();
    
    var thrown = false;
    try {
        doc3.register(name);    
    } catch (e) {
        thrown = true;
    }
    
    assert_false(thrown, 'Exception should not be thrown');
    
}, 'Registry of the import should share master document registry only');

</script>
</body>
</html>
